+++ src/libutil/environment-variables.hh
@@ -29,6 +29,14 @@ std::optional<std::string> getEnvNonEmpty(const std::string & key);
 std::map<std::string, std::string> getEnv();
 
 /**
+ * Like POSIX `setenv`, but always overrides.
+ *
+ * We don't need the non-overriding version, and this is easier to
+ * reimplement on Windows.
+ */
+int setEnv(const char * name, const char * value);
+
+/**
  * Clear the environment.
  */
 void clearEnv();
